feat(user, tasks): Adds crud to tasks with user restrictions and relationships

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { DeleteResult } from 'typeorm';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,12 +15,12 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<DeleteResult> {
-        const removedTasks: DeleteResult = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<DeleteResult> {
+        const removedTasks: DeleteResult = await this.taskRepository.delete({ id, userId: user.id });
 
         if (removedTasks.affected === 0) {
             throw new NotFoundException(`Task with id: ${id} not found.`);
@@ -28,12 +29,12 @@ export class TasksService {
         return removedTasks;
     }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return await this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return await this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const foundTask: Task = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const foundTask: Task = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
         if (!foundTask) {
             throw new NotFoundException(`Task with id: ${id} not found.`);
@@ -42,8 +43,12 @@ export class TasksService {
         return foundTask;
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task: Task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User
+    ): Promise<Task> {
+        const task: Task = await this.getTaskById(id, user);
 
         task.status = status;
